+++ packages/frontend-sdk/src/sdk.ts
@@ -1,6 +1,8 @@
 import {
   DepositManager,
   DepositManager__factory,
+  Handler,
+  Handler__factory,
 } from "@nocturne-xyz/contracts";
 import { WasmJoinSplitProver } from "@nocturne-xyz/local-prover";
 import {
@@ -32,6 +34,8 @@ import {
   OperationStatusResponse,
   toSubmittableOperation,
   decomposeCompressedPoint,
+  SyncOpts,
+  ClosableAsyncIterator,
 } from "@nocturne-xyz/sdk";
 import * as JSON from "bigint-json-serialization";
 import { ContractTransaction } from "ethers";
@@ -58,16 +62,33 @@ export interface Endpoints {
   bundlerEndpoint: string;
 }
 
+export interface ContractAddresses {
+  depositManagerAddress: string;
+  handlerAddress: string;
+}
+
+export interface SyncProgress {
+  latestSyncedMerkleIndex: number;
+}
+
+export interface SyncWithProgressOutput {
+  latestSyncedMerkleIndex: number;
+  latestMerkleIndexOnChain: number;
+  progressIter: ClosableAsyncIterator<SyncProgress>;
+}
+
 export class NocturneFrontendSDK {
   joinSplitProver: WasmJoinSplitProver;
   config: NocturneConfig;
   depositManagerContract: DepositManager;
+  handlerContract: Handler;
   bundlerEndpoint: string;
   screenerEndpoint: string;
 
   private constructor(
     config: NocturneConfig,
     depositManagerContract: DepositManager,
+    handlerContract: Handler,
     endpoints: Endpoints,
     wasmPath: string,
     zkeyPath: string,
@@ -75,9 +96,10 @@ export class NocturneFrontendSDK {
   ) {
     this.config = config;
     this.depositManagerContract = depositManagerContract;
-    this.joinSplitProver = new WasmJoinSplitProver(wasmPath, zkeyPath, vkey);
+    this.handlerContract = handlerContract;
     this.screenerEndpoint = endpoints.screenerEndpoint;
     this.bundlerEndpoint = endpoints.bundlerEndpoint;
+    this.joinSplitProver = new WasmJoinSplitProver(wasmPath, zkeyPath, vkey);
   }
 
   /**
@@ -104,9 +126,13 @@ export class NocturneFrontendSDK {
       signer
     );
 
+    const handlerAddress = config.handlerAddress();
+    const handlerContract = Handler__factory.connect(handlerAddress, signer);
+
     return new NocturneFrontendSDK(
       config,
       depositManagerContract,
+      handlerContract,
       endpoints,
       wasmPath,
       zkeyPath,
@@ -474,18 +500,99 @@ export class NocturneFrontendSDK {
   }
 
   /**
-   * Invoke snap `syncNotes` method.
+   * Start syncing process, returning current merkle index at tip of chain and iterator
+   * returning newly synced merkle indices as syncing process occurs.
    */
-  async sync(): Promise<void> {
-    await window.ethereum.request({
+  async syncWithProgress(syncOpts: SyncOpts): Promise<SyncWithProgressOutput> {
+    let latestMerkleIndexOnChain =
+      (await this.handlerContract.totalCount()).toNumber() - 1;
+    let latestSyncedMerkleIndex =
+      (await this.getLatestSyncedMerkleIndex()) ?? 0;
+
+    const NUM_REFETCHES = 5;
+    const refetchEvery = Math.floor(
+      (latestMerkleIndexOnChain - latestSyncedMerkleIndex) / NUM_REFETCHES
+    );
+
+    let closed = false;
+    const generator = async function* (sdk: NocturneFrontendSDK) {
+      let count = 0;
+      while (!closed && latestSyncedMerkleIndex < latestMerkleIndexOnChain) {
+        latestSyncedMerkleIndex = (await sdk.sync(syncOpts)) ?? 0;
+
+        if (count % refetchEvery === 0) {
+          latestMerkleIndexOnChain =
+            (await sdk.handlerContract.totalCount()).toNumber() - 1;
+        }
+        count++;
+        yield {
+          latestSyncedMerkleIndex,
+        };
+      }
+    };
+
+    const progressIter = new ClosableAsyncIterator(
+      generator(this),
+      async () => {
+        closed = true;
+      }
+    );
+
+    return {
+      latestSyncedMerkleIndex,
+      latestMerkleIndexOnChain,
+      progressIter,
+    };
+  }
+
+  /**
+   * Invoke snap `syncNotes` method, returning latest synced merkle index.
+   */
+  async sync(syncOpts?: SyncOpts): Promise<number | undefined> {
+    const latestSyncedMerkleIndexJson = (await window.ethereum.request({
       method: "wallet_invokeSnap",
       params: {
         snapId: SNAP_ID,
         request: {
           method: "nocturne_sync",
+          params: {
+            syncOpts: syncOpts ? JSON.stringify(syncOpts) : undefined,
+          },
         },
       },
-    });
+    })) as string;
+
+    const latestSyncedMerkleIndex = latestSyncedMerkleIndexJson
+      ? JSON.parse(latestSyncedMerkleIndexJson)
+      : undefined;
+
+    console.log(
+      "[sync] FE-SDK latestSyncedMerkleIndex",
+      latestSyncedMerkleIndex
+    );
+    return latestSyncedMerkleIndex;
+  }
+
+  async getLatestSyncedMerkleIndex(): Promise<number | undefined> {
+    const latestSyncedMerkleIndexJson = (await window.ethereum.request({
+      method: "wallet_invokeSnap",
+      params: {
+        snapId: SNAP_ID,
+        request: {
+          method: "nocturne_getLatestSyncedMerkleIndex",
+        },
+      },
+    })) as string;
+
+    const latestSyncedMerkleIndex = latestSyncedMerkleIndexJson
+      ? JSON.parse(latestSyncedMerkleIndexJson)
+      : undefined;
+
+    console.log(
+      "[getLatestSyncedMerkleIndex] FE-SDK latestSyncedMerkleIndex",
+      latestSyncedMerkleIndex
+    );
+    return latestSyncedMerkleIndex;
   }
 
   /**