+++ packages/sdk/src/sync/rpc/adapter.ts
@@ -18,6 +18,7 @@ import {
 } from "../../utils";
 import {
   fetchJoinSplits,
+  fetchLatestFillBatchEndIndexInRange,
   fetchNotesFromRefunds,
   fetchSubtreeUpdateCommits,
   JoinSplitEvent,
@@ -61,7 +62,7 @@ export class RPCSDKSyncAdapter implements SDKSyncAdapter {
       //
       // We need to filter because we don't know where the start/end total entity indices will be within a block
       const merkleCount = (await handlerContract.count()).toNumber();
-      let lastCommittedMerkleIndex =
+      let latestCommittedMerkleIndex =
         merkleCount !== 0 ? merkleCount - 1 : undefined;
 
       let from = Number(
@@ -88,16 +89,21 @@ export class RPCSDKSyncAdapter implements SDKSyncAdapter {
         await maybeApplyThrottle(to);
 
         // get all events JoinSplit, Refund, and SubtreeUpdate events in the range [`from`, `to`]
-        const [includedNotes, joinSplitEvents, subtreeUpdateCommits] =
-          await Promise.all([
-            fetchNotesFromRefunds(handlerContract, from, to),
-            fetchJoinSplits(handlerContract, from, to),
-            fetchSubtreeUpdateCommits(handlerContract, from, to),
-          ]);
+        const [
+          includedNotes,
+          joinSplitEvents,
+          subtreeUpdateCommits,
+          latestFillBatchEndIndexInRange,
+        ] = await Promise.all([
+          fetchNotesFromRefunds(handlerContract, from, to),
+          fetchJoinSplits(handlerContract, from, to),
+          fetchSubtreeUpdateCommits(handlerContract, from, to),
+          fetchLatestFillBatchEndIndexInRange(handlerContract, from, to),
+        ]);
 
-        // update `lastCommittedMerkleIndex` according to the `SubtreeUpdate` events received
+        // update `latestCommittedMerkleIndex` according to the `SubtreeUpdate` events received
         if (subtreeUpdateCommits.length > 0) {
-          lastCommittedMerkleIndex = maxArray(
+          latestCommittedMerkleIndex = maxArray(
             subtreeUpdateCommits.map(({ inner: { subtreeBatchOffset } }) => {
               return batchOffsetToLatestMerkleIndexInBatch(subtreeBatchOffset);
             })
@@ -131,6 +137,26 @@ export class RPCSDKSyncAdapter implements SDKSyncAdapter {
             (!endTotalEntityIndex || totalEntityIndex <= endTotalEntityIndex)
         );
 
+        const latestMerkleIndexFromNotes =
+          filteredNotes.length > 0
+            ? maxArray(
+                Array.from(filteredNotes.map((n) => n.inner.merkleIndex))
+              )
+            : undefined;
+
+        let latestNewlySyncedMerkleIndex: number | undefined;
+        if (latestFillBatchEndIndexInRange === undefined) {
+          latestNewlySyncedMerkleIndex = latestMerkleIndexFromNotes;
+        } else if (latestMerkleIndexFromNotes === undefined) {
+          latestNewlySyncedMerkleIndex = latestFillBatchEndIndexInRange;
+        } else {
+          // both are defined
+          latestNewlySyncedMerkleIndex = max(
+            latestFillBatchEndIndexInRange,
+            latestMerkleIndexFromNotes
+          );
+        }
+
         currTotalEntityIndex = TotalEntityIndexTrait.fromComponents({
           blockNumber: BigInt(to),
         });
@@ -141,7 +167,8 @@ export class RPCSDKSyncAdapter implements SDKSyncAdapter {
             notes: filteredNotes,
             nullifiers: filteredNullifiers.map((n) => n.inner),
             totalEntityIndex: currTotalEntityIndex,
-            lastCommittedMerkleIndex,
+            latestCommittedMerkleIndex,
+            latestNewlySyncedMerkleIndex,
           };
           yield diff;
         }