+++ packages/snap/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 - change spend key derivation to `keccak256(bip44Node.privateKey)`
+- `nocturne_sync` takes and propagates `SyncOpts` as params
 - remove use of winston logger
 - fix bug where snap was parsing op metadata as `.metadata` not `.opMetadata`
 - setup browser logger for SDK to use